+++ react/tic-tac-toe/src/app/screens/Login/components/CustomInput/styles.module.scss
@@ -1,14 +1,3 @@
-.login-button {
-    background-color: #00ADEE;
-    color: white;
-    width: 150px;
-    height: 50px;
-    align-self: center;
-    border-radius: 5px;
-    margin-top: 10px;
-    font-size: 20px;
-}
-
 .custom-input {
     position: relative;
     margin: 5px 10px 0px 5px;
@@ -24,21 +13,12 @@
   margin-bottom: 10px;
 }
 
-.form {
-    max-width: 360px;
-    padding: 30px;
-    box-shadow: 5px 4px 10px grey;
-    margin: auto;
-    display: flex;
-    flex-direction: column;
-  
+.valid {
+    border-color: green;
 }
-
-.title {
-    align-self: center;
-    color: black;
-    font-size: 30px;
-    margin-bottom: 15px;
+  
+.invalid {
+    border-color: red;
 }
 
 .error {
@@ -47,10 +27,8 @@
     margin-top: 10px;
     margin-left: 10px;
     position: relative;
-    //width: 200px;
     text-align: center;
     border-radius: 3px;
-
 
     &::before {
         content: '';
@@ -63,12 +41,4 @@
         left: 20px;
         top: -8px;
     }
-}
-
-.valid {
-    border-color: green;
-}
-  
-.invalid {
-    border-color: red;
 }
