+++ react/tic-tac-toe/src/app/screens/Login/components/LoginForm/styles.module.scss
@@ -1,0 +1,30 @@
+.login-button {
+    background-color: #00ADEE;
+    color: white;
+    width: 150px;
+    height: 50px;
+    align-self: center;
+    border-radius: 5px;
+    margin-top: 10px;
+    font-size: 20px;
+}
+
+.form {
+    max-width: 360px;
+    padding: 30px;
+    box-shadow: 5px 4px 10px grey;
+    margin: auto;
+    display: flex;
+    flex-direction: column;
+  
+}
+
+.title {
+    align-self: center;
+    color: black;
+    font-size: 30px;
+    margin-bottom: 15px;
+}
+
+
+
